absorb navbar->gamebar->footer
