+++ packages/tictactoe/src/components/GameFooter.tsx
@@ -2,6 +2,11 @@ import React from "react";
 import { Button } from 'reactstrap';
 import { Result, Imperative } from '../core/results';
 import Navbar from "reactstrap/lib/Navbar";
+import { YourMarker, TheirMarker } from './Marker';
+import { Marker } from '../core';
+import hexToBN from "../utils/hexToBN";
+import web3Utils from "web3-utils";
+
 
 interface Props {
   createBlockchainChallenge: () => void;
@@ -10,6 +15,14 @@ interface Props {
   canChallenge: boolean;
   challengeOngoing: boolean;
   result: Result | Imperative;
+  myName: string;
+  opponentName: string;
+  myBalance: string;
+  opponentBalance: string;
+  roundBuyIn: string;
+  you: Marker;
+  rulesRequest: () => void;
+  logoutRequest: () => void;
 }
 
 export default class GameFooter extends React.PureComponent<Props> {
@@ -27,58 +40,136 @@ export default class GameFooter extends React.PureComponent<Props> {
       </Button>
     );
   }
-  renderResultAndButtons(result: Result | Imperative, resign, createBlockchainChallenge, isNotOurTurn, canChallenge) {
-    if (result === Result.YouWin) {
+  myDot = () => (
+    <svg
+      width="20"
+      height="20"
+      viewBox="0 0 20 20"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <circle cx="10" cy="10" r="5" fill="white" />
+      <circle cx="10" cy="10" r="9.5" stroke="white" />
+    </svg>
+  )
+
+  opponentDot = () => (
+    <svg
+      width="10"
+      height="10"
+      viewBox="0 0 10 10"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <circle opacity="0.4" cx="5" cy="5" r="5" fill="white" />
+    </svg>
+  )
+
+  renderMyDots = count =>
+    [...Array(count)].map((e, i) => (
+      <div className="col score-dot" key={i}>
+        {this.myDot()}
+      </div>
+    ))
+
+  renderOpponentDots = count =>
+    [...Array(count)].map((e, i) => (
+      <div className="col score-dot" key={i}>
+        {this.opponentDot()}
+      </div>
+    ))
+
+    renderResultAndButtonsAndDots(
+      result: Result | Imperative,
+      resign, 
+      createBlockchainChallenge,
+      isNotOurTurn,
+      canChallenge,
+      myName,
+      myBalance,
+      myGameCount,
+      opponentGameCount,
+      opponentBalance,
+      opponentName) {
       return (
+        <div>
+        <Navbar className="game-bar fixed-top">
+        <Button color="link" className="navbar-button mr-auto" onClick={this.props.rulesRequest}>
+          Rules
+        </Button>
+        <div className="container">
+          <div className="col-2 my-name text-right"><YourMarker you={this.props.you} />{myName}</div>
+          <div className="col-auto my-balance">
+            <div className="text-center">
+              {web3Utils.fromWei(myBalance, "ether")}
+            </div>
+            <div className="eth text-center">ETH</div>
+          </div>
+          <div className="col">
+            <div className="row">
+              {this.renderMyDots(myGameCount)}
+              {this.renderOpponentDots(opponentGameCount)}
+            </div>
+          </div>
+          <div className="col-auto opponent-balance">
+            <div className="text-center">
+              {web3Utils.fromWei(opponentBalance, "ether")}
+            </div>
+            <div className="eth text-center">ETH</div>
+          </div>
+          <span className="col-2 opponent-name">{opponentName}<TheirMarker you={this.props.you} /></span>
+        </div>
+        
+        <Button color="link" className="navbar-button ml-auto" onClick={this.props.logoutRequest}>
+          Sign Out
+        </Button>
+      </Navbar>
         <Navbar id="you-win" className="navbar fixed-bottom footer-bar">
           {this.renderResignButton(resign, isNotOurTurn)}
           <span>You Win!</span>
           {this.renderChallengeButton( createBlockchainChallenge, canChallenge)}
         </Navbar>
+        </div>
       );
     }
-    if (result === Result.YouLose) {
-      return (
-        <Navbar id="you-lose" className="navbar fixed-bottom footer-bar">
-          {this.renderResignButton(resign, isNotOurTurn)}
-          <span>You Lose!</span>
-          {this.renderChallengeButton( createBlockchainChallenge, canChallenge)}
-        </Navbar>
-      );
-    }
-    if (result === Result.Tie) {
-      return (
-        <Navbar id="tie" className="navbar fixed-bottom footer-bar">
-          {this.renderResignButton(resign, isNotOurTurn)}
-          <span>It's a Draw!</span>
-          {this.renderChallengeButton( createBlockchainChallenge, canChallenge)}
-        </Navbar>
-      );
-    }
-    if (result === Imperative.Choose) {
-      return (
-        <Navbar id="choose" className="navbar fixed-bottom footer-bar">
-        {this.renderResignButton(resign, isNotOurTurn)}
-        <span>Choose your move</span>
-        {this.renderChallengeButton( createBlockchainChallenge, canChallenge)}
-      </Navbar>
-      );
-    }
-    if (result === Imperative.Wait) {
-      return (
-        <Navbar id="wait" className="navbar fixed-bottom footer-bar">
-          {this.renderResignButton(resign, isNotOurTurn)}
-          <span>Wait for Opponent's move!</span>
-          {this.renderChallengeButton( createBlockchainChallenge, canChallenge)}
-        </Navbar>
-      );
-    } else { return <span>&nbsp;</span>; }
-  }
+      
   render() {
-    const { resign, createBlockchainChallenge, isNotOurTurn, canChallenge, result } = this.props;
-
-    return (
-      this.renderResultAndButtons(result, resign, createBlockchainChallenge, isNotOurTurn, canChallenge)
+    const {
+      myName,
+      opponentName,
+      myBalance,
+      opponentBalance,
+      roundBuyIn,
+      result,
+      resign,
+      createBlockchainChallenge,
+      isNotOurTurn,
+      canChallenge,
+    } = this.props;
+    
+    const myGameCount = Math.round(
+      hexToBN(myBalance)
+        .div(hexToBN(roundBuyIn))
+        .toNumber()
+    );
+    const opponentGameCount = Math.round(
+      hexToBN(opponentBalance)
+        .div(hexToBN(roundBuyIn))
+        .toNumber()
+    );
+      return (
+      this.renderResultAndButtonsAndDots(
+        result,
+        resign, 
+        createBlockchainChallenge,
+        isNotOurTurn,
+        canChallenge,
+        myName,
+        myBalance,
+        myGameCount,
+        opponentGameCount,
+        opponentBalance,
+        opponentName)
     );
   }
 }
