+++ packages/tictactoe/src/components/GameScreenNoGameBar.tsx
@@ -1,11 +1,11 @@
 import React from "react";
-import NavigationBarContainer from "../containers/NavigationBarContainer";
+// import NavigationBarContainer from "../containers/NavigationBarContainer";
 import GameFooterContainer from "../containers/GameFooterContainer";
 
 export const GameScreenNoGameBar = props => {
   return (
     <div className="w-100">
-      <NavigationBarContainer />
+      {/* <NavigationBarContainer /> */}
       <div className="container centered-container w-100 game-container">
         {props.children}
       </div>
