--- packages/tictactoe/src/containers/GameBarContainer.tsx
@@ -1,29 +1,0 @@
-import { connect } from 'react-redux';
-
-import GameBar from '../components/GameBar';
-import { SiteState } from '../redux/reducer';
-import { PlayingState } from '../redux/game/state';
-
-function mapStateToProps(state: SiteState) {
-  const gameState = state.game.gameState as PlayingState;
-  const { myName, opponentName, roundBuyIn, player, onScreenBalances} = gameState;
-
-  const myBalance = onScreenBalances[player];
-  const opponentBalance = onScreenBalances[1 - player];
-
-  return {
-    myName,
-    opponentName,
-    myBalance,
-    opponentBalance,
-    roundBuyIn,
-  };
-}
-
-const mapDispatchToProps = {
-};
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(GameBar);
