+++ packages/tictactoe/src/containers/GameFooterContainer.tsx
@@ -13,6 +13,8 @@ import {
   OsWaitForOpponentToPickMove
 } from "../redux/game/state";
 import { Player } from "../core/players";
+import * as loginActions from "../redux/login/actions";
+import * as globalActions from "../redux/global/actions";
 
 function mapStateToProps(state: SiteState) {
   const gameState = state.game.gameState as
@@ -32,16 +34,33 @@ function mapStateToProps(state: SiteState) {
   const challengeOngoing =
     gameState.name === "OS_PICK_CHALLENGE_MOVE" ||
     gameState.name === "XS_PICK_CHALLENGE_MOVE";
+    const { myName, opponentName, roundBuyIn, onScreenBalances} = gameState;
+  
+    const myBalance = onScreenBalances[player];
+    const opponentBalance = onScreenBalances[1 - player];
+    const name =
+    "myName" in state.game.gameState ? state.game.gameState.myName : "";
+  const you = "you" in state.game.gameState ? state.game.gameState.you : "";
   return {
     isNotOurTurn,
     canChallenge,
     challengeOngoing,
     result,
+    myName,
+    opponentName,
+    myBalance,
+    opponentBalance,
+    roundBuyIn,
+    showRules: state.rules.visible,
+    loginDisplayName: name,
+    you,
   };
 }
 const mapDispatchToProps = {
   resign: gameActions.resign,
   createBlockchainChallenge: gameActions.createChallenge,
+  logoutRequest: loginActions.logoutRequest,
+  rulesRequest: globalActions.toggleVisibility,
 };
 
 export default connect(
