+++ packages/tictactoe/src/styles/footer.scss
@@ -16,7 +16,8 @@
   position: absolute;
   top: 50%;
   left: 50%;
-  transform: translate(-50%,-50%) translate(0,$navbar-height);
+  transform: translate(-50%,-50%);
+  // translate(0,$navbar-height);
 }
 
 .footer-resign:hover, .footer-challenge:hover, .footer-playagain:hover {
