+++ packages/tictactoe/src/styles/game-bar.scss
@@ -9,7 +9,7 @@
 
 .game-bar .my-name, .game-bar .opponent-name {
   font-family: $ttt-font-family;
-  font-size: 16px;
+  font-size: 24px;
   font-weight: bold;
   font-style: normal;
   font-stretch: normal;
