+++ packages/tictactoe/src/styles/navbar.scss
@@ -44,8 +44,8 @@
   }
   
   .status {
-    background-color: $ttt-background;
-    font-size: 50px;
+    // background-color: $ttt-background;
+    // font-size: 50px;
     text-emphasis: bold;
     font-family: monospace;
   }
